# docs/swagger.yaml
components:
  schemas:
...
    BaseResponse:
      properties:
        msg:
          default: ''
          title: Msg
          type: string
        status:
          default: success
          title: Status
          type: string
      title: BaseResponse
      type: object
...
    CreateUser:
      properties:
        email:
          format: email
          title: Email
          type: string
        first_name:
          title: First Name
          type: string
        is_user_agreement_accepted:
          default: false
          title: Is User Agreement Accepted
          type: boolean
        last_name:
          title: Last Name
          type: string
        position:
          anyOf:
          - type: string
          - type: 'null'
          title: Position
        second_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Second Name
        username:
          title: Username
          type: string
      required:
      - email
      - username
      - first_name
      - last_name
      title: CreateUser
      type: object
...
    FrontendUserLogin:
      properties:
        email:
          anyOf:
          - format: email
            type: string
          - type: 'null'
          title: Email
        username:
          anyOf:
          - type: string
          - type: 'null'
          title: Username
      title: FrontendUserLogin
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    UserSignResponse:
      properties:
        email_to:
          format: email
          title: Email To
          type: string
        expire_minutes:
          default: 5
          title: Expire Minutes
          type: integer
        verify_token:
          title: Verify Token
          type: string
      required:
      - email_to
      - verify_token
      title: UserSignResponse
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    WikiAccessToken:
      in: cookie
      name: wiki_access_token
      type: apiKey
    WikiApiKey:
      in: header
      name: X-Wiki-API-Key
      type: apiKey
    WikiBearer:
      scheme: bearer
      type: http
    WikiRootApiKey:
      description: 'Root key, allows you to perform actions as administrator, random
        user data is generated.

        **Limitation:** operations requiring an explicitly registered userare unavailable,
        example.: `Create: POST /api_key/`.'
      in: header
      name: X-Wiki-ROOT-API-Key
      type: apiKey
info:
  description: Welcome to Wiki's API documentation! Here you will able to discover
    all of the ways you can interact with the Wiki API.
  title: Wiki API
  version: 0.0.1
openapi: 3.1.0
paths:
...
  /api/v1/auth/login:
    post:
      description: '## Logging in to a user account

        The endpoint for sending an email or username and receiving a login confirmation
        token.

        A confirmation code will be sent to the specified email.'
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FrontendUserLogin'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSignResponse'
          description: Verification token data.
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Login by email or username
      tags:
      - Auth
  /api/v1/auth/signup:
    post:
      description: '## Registration - submitting an application

        Sending an application for registration in the system.'
      operationId: signup_api_v1_auth_signup_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSignResponse'
          description: Verification token data.
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Signup with user data entry
      tags:
      - Auth
  /api/v1/auth/verify:
    get:
      description: '## Verification on the sent code

        Endpoint for confirming the code received by mail, issuing access token to
        perform authorized actions.'
      operationId: verify_api_v1_auth_verify_get
      parameters:
      - in: query
        name: token
        required: true
        schema:
          title: Token
          type: string
      - in: query
        name: verification_code
        required: true
        schema:
          title: Verification Code
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
          description: Cookies are set. Success message.
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Verification by token and code
      tags:
      - Auth
...
servers:
- description: Local server
  url: http://localhost:8000/
- description: Demo server
  url: https://api.wiki.example.com/
- description: Ngrok free server
  url: https://poorly-ideal-cobra.ngrok-free.app/